pyproject.toml
[project]
name = "detr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bidict>=0.23.1",
    "dotenv>=0.9.9",
    "einops>=0.8.1",
    "hydra-core>=1.3.2",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "pytz>=2025.2",
    "scipy>=1.17.0",
    "tabulate>=0.9.0",
    "torch==2.9.1",
    "torchscan>=0.1.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "wandb>=0.24.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu129", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu129", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
